CIS Portfolio @ University of Louisville, College of Business
This is a collection of my time here at the University of Louisville, in the College of Business studying Computer Information Systems with a concentration in Information Security and Business Process Management. Here you will find work from my current and past classes detailing my work in all of these aspects as an individual and as a collaborative group member. Feel free to contact me via my resume which is also liked in this repository.